workers/dtcc-core/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colorlog==6.9.0
colormap==1.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dtcc==0.7.21
dtcc-builder==0.7.21
dtcc-common==0.7.21
dtcc-data==0.7.21
dtcc-io==0.7.21
dtcc-model==0.7.21
dtcc-viewer==0.7.21
dtcc-wrangler==0.7.21
easydev==0.13.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
fiona==1.10.1
Flask==3.1.0
folium==0.19.4
fonttools==4.55.3
fqdn==1.5.1
glfw==2.8.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imgui==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
laspy==2.5.4
lazrs==0.6.2
line_profiler==4.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.24.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meshio==5.3.5
mistune==3.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.1
psycopg==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyassimp==5.2.5
pybind11==2.10.4
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyglet==2.0.20
pyglfw==0.2.2
pygltflib==1.16.3
Pygments==2.19.1
PyOpenGL==3.1.7
pyparsing==3.2.1
pypoints2grid==0.1.9
pyproj==3.7.0
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
PyYAML==6.0.2
pyzmq==26.2.0
rasterio==1.4.3
rasterstats==0.20.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.0
Send2Trash==1.8.3
setuptools==75.8.0
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xyzservices==2024.9.0
GDAL==3.6.2
